float32

+++ projects/ros2can/src/telemetry_demux.py
@@ -19,8 +19,8 @@ class TelemetryDemuxNode(Node):
             10) # QoS profile depth
 
         # Create clean, dedicated publishers for Foxglove
-        self.speed_pub = self.create_publisher(Int32, '/rear_ecu/speedometer', 10)
-        self.tach_pub = self.create_publisher(Int32, '/rear_ecu/tachometer', 10)
+        self.speed_pub = self.create_publisher(Float32, '/rear_ecu/speedometer', 10)
+        self.tach_pub = self.create_publisher(Float32, '/rear_ecu/tachometer', 10)
 
         self.get_logger().info("Rear Ecu Demuxer Started. Listening to /rear_ecu/rpdo...")
 
@@ -34,8 +34,8 @@ class TelemetryDemuxNode(Node):
 
         # Index 0x2001 (Decimal 8193) -> Tachometer
         elif msg.index == 8193:
-            tach_msg = Int32()
-            tach_msg.data = int(msg.data)
+            tach_msg = Float32()
+            tach_msg.data = float(msg.data)
             self.tach_pub.publish(tach_msg)
 
 def main(args=None):
